# backend/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Estevas Garden API",
    "version": "1.1.0",
    "description": "REST API for managing customers and products in the Estevas Garden platform."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://estevas-garden-project.onrender.com/",
      "description": "Production server"
    }
  ],
  "paths": {
    "/auth/profile": {
      "get": {
        "summary": "Retorna as informações do usuário autenticado",
        "tags": ["Authentication"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Usuário autenticado" },
          "401": { "description": "Token inválido ou ausente" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterUser" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Email already exists or invalid input" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login and receive a JWT",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginUser" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, token returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "description": "Invalid email or password" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Login successful" },
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        }
      },
      "LoginUser": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "example": "user@example.com" },
          "password": { "type": "string", "example": "mypassword123" }
        }
      },
      "RegisterUser": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "example": "Ana Silva" },
          "email": { "type": "string", "example": "ana@example.com" },
          "password": { "type": "string", "example": "mypassword123" }
        }
      },
...
    }
  }
}